Use client function get_blockchain_info
Currently we are using the low level `call` function to get at the IBD.
The client provides a higher level function to call the
`getblockchaininfo` RPC method.

The RPC method returns a type `GetBlockchainInfo`. As for all methods
this type is version specific. Calling `into_model` converts this type
to the version in-specific `model::GetBlockchainInfo` type. This type is
forward compatible - i.e., `corepc-types` guarantees there will always
be an `into_model` function and that if the fields change downstream
will fail at build time not runtime.